Refactors code out into separate files
